move normalize condition be a util function
